feat: Add result property to CryptoStore
This commit adds a new property called `result` to the `CryptoStore` type in the `store.ts` file. The `result` property is of type `CryptoPriceSchema` and is used to store the fetched data from the API, including the image URL, last update time, price, high day value, low day value, and 24-hour change percentage. This enhancement improves the functionality of the application by providing access to the latest cryptocurrency price data.